+++ packages/ui/core-components/.storybook/preview.js
@@ -5,7 +5,6 @@ import { initialize } from '../src/lib/storybook-helpers/initializeUSQL.js';
 /** @type { import('@storybook/svelte').Preview } */
 const preview = {
 	parameters: {
-		actions: { argTypesRegex: '^on[A-Z].*' },
 		controls: {
 			matchers: {
 				color: /(background|color)$/i,