+++ packages/ui/core-components/src/lib/unsorted/viz/references/ReferencePoint/ReferencePoint.stories.svelte
@@ -118,6 +118,20 @@
 
 	const inputStore = writable({});
 	setContext(INPUTS_CONTEXT_KEY, inputStore);
+
+	const data = Query.create(
+		`
+		select *
+		from numeric_series
+		where series in (
+			select series
+			from numeric_series
+			order by series asc
+			limit 1
+		)
+		`,
+		query
+	);
 </script>
 
 <Story
@@ -126,7 +140,6 @@
 	argTypes={{ x: { control: 'number' }, y: { control: 'number' } }}
 	let:args
 >
-	{@const data = Query.create(`SELECT * FROM numeric_series WHERE series='pink'`, query)}
 	<LineChart x="x" y="y" {data}>
 		<ReferencePoint {...args} />
 	</LineChart>
@@ -138,23 +151,26 @@
 	argTypes={{ x: { control: 'text' }, y: { control: 'text' } }}
 	let:args
 >
-	{@const chartData = Query.create(`SELECT * FROM numeric_series WHERE series='pink'`, query)}
 	{@const referencePointData = Query.create(
 		`
-		SELECT
+		select
 			x,
 			y,
 			row_number() over(order by x) as label
-		FROM numeric_series
-		WHERE
-			series='pink' AND
-			x in (30, 50, 70)
-	`,
+		from numeric_series
+		where
+			series in (
+				select series
+				from numeric_series
+				order by series asc
+				limit 1
+			)
+			and x in (30, 50, 70)
+		`,
 		query
 	)}
-
-	<QueryLoad data={chartData}>
-		<LineChart x="x" y="y" data={chartData}>
+	<QueryLoad {data}>
+		<LineChart x="x" y="y" {data}>
 			<QueryLoad data={referencePointData}>
 				<ReferencePoint {...args} data={referencePointData} />
 			</QueryLoad>
@@ -168,7 +184,6 @@
 	argTypes={{ label: { control: false } }}
 	let:args
 >
-	{@const data = Query.create(`SELECT * FROM numeric_series WHERE series='pink'`, query)}
 	<LineChart x="x" y="y" {data}>
 		<ReferencePoint {...args}>This label is passed via the default slot</ReferencePoint>
 	</LineChart>
@@ -189,7 +204,6 @@
 </Story>
 
 <Story name="Colors">
-	{@const data = Query.create(`SELECT * FROM numeric_series WHERE series='pink'`, query)}
 	<LineChart x="x" y="y" {data}>
 		<ReferencePoint x="10" y="100" color="blue" label="blue" />
 		<ReferencePoint x="20" y="100" color="red" label="red" />
@@ -202,14 +216,12 @@
 
 <!-- Specifying x without y -->
 <Story name="Error" args={{ x: 24 }} let:args>
-	{@const data = Query.create(`SELECT * FROM numeric_series WHERE series='pink'`, query)}
 	<LineChart x="x" y="y" {data}>
 		<ReferencePoint {...args} />
 	</LineChart>
 </Story>
 
 <Story id="callout" name="Callout" args={{ x: 24, y: 514, label: 'This is a Callout!' }} let:args>
-	{@const data = Query.create(`SELECT * FROM numeric_series WHERE series='pink'`, query)}
 	<LineChart x="x" y="y" {data}>
 		<Callout {...args} />
 	</LineChart>
@@ -220,7 +232,6 @@
 	args={{ x: 24, y: 514, labelWidth: 'fit', label: 'This is a Callout!' }}
 	let:args
 >
-	{@const data = Query.create(`SELECT * FROM numeric_series WHERE series='pink'`, query)}
 	<LineChart x="x" y="y" {data}>
 		<Callout {...args} />
 	</LineChart>
@@ -232,7 +243,6 @@
 	argTypes={{ label: { control: false } }}
 	let:args
 >
-	{@const data = Query.create(`SELECT * FROM numeric_series WHERE series='pink'`, query)}
 	<LineChart x="x" y="y" {data}>
 		<!-- prettier-ignore -->
 		<Callout {...args}>
@@ -267,7 +277,6 @@
 	}}
 	let:args
 >
-	{@const data = Query.create(`SELECT * FROM numeric_series WHERE series='pink'`, query)}
 	<LineChart x="x" y="y" {data}>
 		<ReferencePoint {...args} />
 	</LineChart>
@@ -283,26 +292,25 @@
 	argTypes={{ x: { control: 'text' }, y: { control: 'text' } }}
 	let:args
 >
-	{@const chartData = Query.create(`SELECT * FROM numeric_series WHERE series='pink'`, query)}
 	{@const referencePointData = Query.create(
 		`
-		SELECT
+		select
 			x,
 			y,
 			row_number() over(order by x) as label
-		FROM numeric_series
-		WHERE
-			series='pink' AND
-			x in (30, 50, 70)
-	`,
+		from numeric_series
+		where
+			series in (
+				select series
+				from numeric_series
+				order by series asc
+				limit 1
+			)
+			and x in (30, 50, 70)
+		`,
 		query
 	)}
-
-	<QueryLoad data={chartData}>
-		<LineChart x="x" y="y" data={chartData}>
-			<QueryLoad data={referencePointData}>
-				<ReferencePoint {...args} data={referencePointData} />
-			</QueryLoad>
-		</LineChart>
-	</QueryLoad>
+	<LineChart x="x" y="y" {data}>
+		<ReferencePoint {...args} data={referencePointData} />
+	</LineChart>
 </Story>